Forgot the END directive is optional and people like to use spaces not tabs...
All processed lines will now be scanned for whitespaces and replaced with tabs. The END directive's first executable address is optional and will default to the START directive memory address if not specified. Removed the print out of the symbol table and SIC source to limit the message box size so a user can click 'OK'..