Reworked a option for loading an object file. Renamed to match intent the intent of Relocation rather than being a duplicate loading option. Moved the logic to the construction of the dialogue box so that the relocating loader can be used elsewhere with less repeated code each time.
Added in extra dialogue box prompting the user for either Absolute loading or Relocating loading. Currently Relocation Loading defaults back to absolute loading and provides a dialogue box indicating success. This is a basis for calling the Relocating Loader once it is implemented.

Modified an existing dialogue box for Relocating. The values are now determined internally from the object lines rather than being calculated and passed in as arguments. Updated a note informing the user of the range available to them and repaired an input field to accept 6 characters of input instead of 4 to match standards.

Added in a Warning Dialogue box to relocation. If you attempt to relocate a program that is already in memory, you are copying that program. Within the dialogue instructions are provided for loading directly to a position as desired.