Test if device ID is stored in BYTE constant
Currently, for instructions that read/write to I/O devices, the logic in the FetchWord method in the SIC_CPU.cs assumes that device ID's are stored in data words. The effect of this is that FetchWord retrieves the next three bytes of data starting from the contents of the Target Address. This is a problem if a SIC file writer decides to store a device ID using the BYTE directive because FetchWord will read two bytes beyond what is necessary and assign an invalid number to the DeviceNumberToWriteTo or DeviceNumberToRead int values, and ultimately cause an "array out of bounds" exception.